pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ktoolbox"
version = "0.8.0"
description = "Utility library"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
]

[project.optional-dependencies]
PyYAML = ["PyYAML"]
jinja2 = ["jinja2"]
paramiko = ["paramiko"]
pyserial = ["pyserial"]
all = ["PyYAML", "jinja2", "paramiko", "pyserial"]

[tool.setuptools.package-data]
ktoolbox = ["py.typed"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests"]

[tool.black]

[tool.pytest.ini_options]
pythonpath = [
    ".",
]

[tool.mypy]
strict = true
files = "."
exclude = "build/"

[[tool.mypy.overrides]]
module = "serial"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "setuptools"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pytest"
ignore_missing_imports = true
